Move page and post forms into the forms package, due to their newfound complexity.
