+++ corvid/main/forms/__init__.py
@@ -1,3 +1,5 @@
 from .registration import RegistrationForm
 from .create_project import CreateProjectForm
 from .clone_project import CloneProjectForm
+from .page import PageForm
+from .post import PostForm
