+++ corvid/main/views/page.py
@@ -4,34 +4,10 @@ from django.views.generic.edit import UpdateView
 from django.template.response import TemplateResponse
 from django.core.urlresolvers import reverse
 from django.shortcuts import get_object_or_404
-from django import forms
-from pagedown.widgets import PagedownWidget
 
-from main.models import Project, Page, User
+from ..models import Project, Page, User
+from ..forms import PageForm
 from .protected_view import ProtectedViewMixin
-
-
-class PageForm(forms.ModelForm):
-
-    def __init__(self, *args, **kwargs):
-        # Order is important here: project must be removed from kwargs or super
-        # constructor fails.  Querysets must be modified after super
-        # constructor finishes, otherwise the self.fields dict is not yet
-        # initialized.
-        project = kwargs.pop('__project')
-        super(PageForm, self).__init__(*args, **kwargs)
-        plugins = project.pageplugin_set.all()
-        self.fields['post_plugins'].queryset = plugins
-
-    class Meta:
-        model = Page
-        fields = ['title', 'post_plugins', 'content']
-        widgets = {
-            'content': PagedownWidget(),
-        }
-        help_texts = {
-            'post_plugins': 'Hold "Control" while clicking to select multiple.'
-        }
 
 
 class PageBase:
